create positions
위치 정보와 지도 부분 분리
